refactor(server): isolate external authentication routes
